Move places handling from translator to MainFratureEmitter.
