+++ generator/osm_source.cpp
@@ -18,6 +18,7 @@
 #include "platform/platform.hpp"
 
 #include "geometry/mercator.hpp"
+#include "geometry/tree4d.hpp"
 
 #include "coding/parse_xml.hpp"
 
@@ -183,6 +184,78 @@ public:
   }
 };
 
+// TODO(mgsergio): comment
+class Place
+{
+  FeatureBuilder1 m_ft;
+  m2::PointD m_pt;
+  uint32_t m_type;
+  double m_thresholdM;
+
+  bool IsPoint() const { return (m_ft.GetGeomType() == feature::GEOM_POINT); }
+  static bool IsEqualTypes(uint32_t t1, uint32_t t2)
+  {
+    // Use 2-arity places comparison for filtering.
+    // ("place-city-capital-2" is equal to "place-city")
+    ftype::TruncValue(t1, 2);
+    ftype::TruncValue(t2, 2);
+    return (t1 == t2);
+  }
+
+public:
+  Place(FeatureBuilder1 const & ft, uint32_t type) : m_ft(ft), m_pt(ft.GetKeyPoint()), m_type(type)
+  {
+    using namespace ftypes;
+
+    switch (IsLocalityChecker::Instance().GetType(m_type))
+    {
+    case COUNTRY: m_thresholdM = 300000.0; break;
+    case STATE: m_thresholdM = 100000.0; break;
+    case CITY: m_thresholdM = 30000.0; break;
+    case TOWN: m_thresholdM = 20000.0; break;
+    case VILLAGE: m_thresholdM = 10000.0; break;
+    default: m_thresholdM = 10000.0; break;
+    }
+  }
+
+  FeatureBuilder1 const & GetFeature() const { return m_ft; }
+
+  m2::RectD GetLimitRect() const
+  {
+    return MercatorBounds::RectByCenterXYAndSizeInMeters(m_pt, m_thresholdM);
+  }
+
+  bool IsEqual(Place const & r) const
+  {
+    return (IsEqualTypes(m_type, r.m_type) &&
+            m_ft.GetName() == r.m_ft.GetName() &&
+            (IsPoint() || r.IsPoint()) &&
+            MercatorBounds::DistanceOnEarth(m_pt, r.m_pt) < m_thresholdM);
+  }
+
+  /// Check whether we need to replace place @r with place @this.
+  bool IsBetterThan(Place const & r) const
+  {
+    // Check ranks.
+    uint8_t const r1 = m_ft.GetRank();
+    uint8_t const r2 = r.m_ft.GetRank();
+    if (r1 != r2)
+      return (r2 < r1);
+
+    // Check types length.
+    // ("place-city-capital-2" is better than "place-city").
+    uint8_t const l1 = ftype::GetLevel(m_type);
+    uint8_t const l2 = ftype::GetLevel(r.m_type);
+    if (l1 != l2)
+      return (l2 < l1);
+
+    // Assume that area places has better priority than point places at the very end ...
+    /// @todo It was usefull when place=XXX type has any area fill style.
+    /// Need to review priority logic here (leave the native osm label).
+    return !IsPoint();
+  }
+};
+
 class MainFeaturesEmitter
 {
   using TWorldGenerator = WorldMapGenerator<feature::FeaturesCollector>;
@@ -195,6 +268,9 @@ class MainFeaturesEmitter
 
   string m_srcCoastsFile;
   bool m_failOnCoasts;
+
+  // TODO(mgsergio): comment.
+  m4::Tree<Place> m_places;
 
   enum TypeIndex
   {
@@ -248,47 +324,33 @@ public:
       m_world.reset(new TWorldGenerator(info));
   }
 
-  void operator()(FeatureBuilder1 fb)
+  void operator()(FeatureBuilder1 & fb)
   {
-    uint32_t const coastType = Type(NATURAL_COASTLINE);
-    bool const hasCoast = fb.HasType(coastType);
+    static uint32_t const placeType = classif().GetTypeByPath({"place"});
+    uint32_t const type = fb.GetParams().FindType(placeType, 1);
 
-    if (m_coasts)
+    if (type != ftype::GetEmptyValue() && !fb.GetName().empty())
     {
-      if (hasCoast)
-      {
-        CHECK(fb.GetGeomType() != feature::GEOM_POINT, ());
-        // leave only coastline type
-        fb.SetType(coastType);
-        (*m_coasts)(fb);
-      }
-      return;
+      m_places.ReplaceEqualInRect(
+          Place(fb, type),
+          [](Place const & p1, Place const & p2) { return p1.IsEqual(p2); },
+          [](Place const & p1, Place const & p2) { return p1.IsBetterThan(p2); });
     }
-
-    if (hasCoast)
+    else
     {
-      fb.PopExactType(Type(NATURAL_LAND));
-      fb.PopExactType(coastType);
+      Emit(fb);
     }
-    else if ((fb.HasType(Type(PLACE_ISLAND)) || fb.HasType(Type(PLACE_ISLET))) &&
-             fb.GetGeomType() == feature::GEOM_AREA)
-    {
-      fb.AddType(Type(NATURAL_LAND));
-    }
-
-    if (!fb.RemoveInvalidTypes())
-      return;
-
-    if (m_world)
-      (*m_world)(fb);
-
-    if (m_countries)
-      (*m_countries)(fb);
   }
 
   /// @return false if coasts are not merged and FLAG_fail_on_coasts is set
   bool Finish()
   {
+    m_places.ForEach([this](Place const & p)
+    {
+      // m_places are no longer used after this point.
+      Emit(const_cast<FeatureBuilder1 &>(p.GetFeature()));
+    });
+
     if (m_world)
       m_world->DoMerge();
 
@@ -346,6 +408,45 @@ public:
       names = m_countries->Parent().Names();
     else
       names.clear();
+  }
+
+private:
+  void Emit(FeatureBuilder1 & fb)
+  {
+    uint32_t const coastType = Type(NATURAL_COASTLINE);
+    bool const hasCoast = fb.HasType(coastType);
+
+    if (m_coasts)
+    {
+      if (hasCoast)
+      {
+        CHECK(fb.GetGeomType() != feature::GEOM_POINT, ());
+        // leave only coastline type
+        fb.SetType(coastType);
+        (*m_coasts)(fb);
+      }
+      return;
+    }
+
+    if (hasCoast)
+    {
+      fb.PopExactType(Type(NATURAL_LAND));
+      fb.PopExactType(coastType);
+    }
+    else if ((fb.HasType(Type(PLACE_ISLAND)) || fb.HasType(Type(PLACE_ISLET))) &&
+             fb.GetGeomType() == feature::GEOM_AREA)
+    {
+      fb.AddType(Type(NATURAL_LAND));
+    }
+
+    if (!fb.RemoveInvalidTypes())
+      return;
+
+    if (m_world)
+      (*m_world)(fb);
+
+    if (m_countries)
+      (*m_countries)(fb);
   }
 };
 }  // anonymous namespace
@@ -521,7 +622,7 @@ bool GenerateFeaturesImpl(feature::GenerateInfo & info)
                                              info.m_bookingReferenceDir);
 
     stringstream skippedElements;
-    
+
     // Here we can add new tags to element!!!
     auto const fn = [&](OsmElement * e)
     {
@@ -534,7 +635,7 @@ bool GenerateFeaturesImpl(feature::GenerateInfo & info)
         skippedElements << e->id << endl;
         return;
       }
-      
+
       parser.EmitElement(e);
     };
 
@@ -567,8 +668,6 @@ bool GenerateFeaturesImpl(feature::GenerateInfo & info)
         LOG(LERROR, ("Can't output into", skippedElementsPath));
       }
     }
-    
-    parser.Finish();
 
     // Stop if coasts are not merged and FLAG_fail_on_coasts is set
     if (!bucketer.Finish())
