+++ generator/osm_translator.hpp
@@ -9,8 +9,6 @@
 #include "indexer/feature_visibility.hpp"
 #include "indexer/ftypes_matcher.hpp"
 
-#include "geometry/tree4d.hpp"
-
 #include "coding/file_writer.hpp"
 
 #include "base/cache.hpp"
@@ -24,77 +22,6 @@
 
 namespace
 {
-class Place
-{
-  FeatureBuilder1 m_ft;
-  m2::PointD m_pt;
-  uint32_t m_type;
-  double m_thresholdM;
-
-  bool IsPoint() const { return (m_ft.GetGeomType() == feature::GEOM_POINT); }
-  static bool IsEqualTypes(uint32_t t1, uint32_t t2)
-  {
-    // Use 2-arity places comparison for filtering.
-    // ("place-city-capital-2" is equal to "place-city")
-    ftype::TruncValue(t1, 2);
-    ftype::TruncValue(t2, 2);
-    return (t1 == t2);
-  }
-
-public:
-  Place(FeatureBuilder1 const & ft, uint32_t type) : m_ft(ft), m_pt(ft.GetKeyPoint()), m_type(type)
-  {
-    using namespace ftypes;
-
-    switch (IsLocalityChecker::Instance().GetType(m_type))
-    {
-    case COUNTRY: m_thresholdM = 300000.0; break;
-    case STATE: m_thresholdM = 100000.0; break;
-    case CITY: m_thresholdM = 30000.0; break;
-    case TOWN: m_thresholdM = 20000.0; break;
-    case VILLAGE: m_thresholdM = 10000.0; break;
-    default: m_thresholdM = 10000.0; break;
-    }
-  }
-
-  FeatureBuilder1 const & GetFeature() const { return m_ft; }
-
-  m2::RectD GetLimitRect() const
-  {
-    return MercatorBounds::RectByCenterXYAndSizeInMeters(m_pt, m_thresholdM);
-  }
-
-  bool IsEqual(Place const & r) const
-  {
-    return (IsEqualTypes(m_type, r.m_type) &&
-            m_ft.GetName() == r.m_ft.GetName() &&
-            (IsPoint() || r.IsPoint()) &&
-            MercatorBounds::DistanceOnEarth(m_pt, r.m_pt) < m_thresholdM);
-  }
-
-  /// Check whether we need to replace place @r with place @this.
-  bool IsBetterThan(Place const & r) const
-  {
-    // Check ranks.
-    uint8_t const r1 = m_ft.GetRank();
-    uint8_t const r2 = r.m_ft.GetRank();
-    if (r1 != r2)
-      return (r2 < r1);
-
-    // Check types length.
-    // ("place-city-capital-2" is better than "place-city").
-    uint8_t const l1 = ftype::GetLevel(m_type);
-    uint8_t const l2 = ftype::GetLevel(r.m_type);
-    if (l1 != l2)
-      return (l2 < l1);
-
-    // Assume that area places has better priority than point places at the very end ...
-    /// @todo It was usefull when place=XXX type has any area fill style.
-    /// Need to review priority logic here (leave the native osm label).
-    return !IsPoint();
-  }
-};
-
 /// Generated features should include parent relation tags to make
 /// full types matching and storing any additional info.
 class RelationTagsBase
@@ -243,7 +170,6 @@ protected:
     }
   }
 };
-
 }  // namespace
 
 /// @param  TEmitter  Feature accumulating policy
@@ -255,7 +181,7 @@ class OsmToFeatureTranslator
   TCache & m_holder;
   uint32_t m_coastType;
   unique_ptr<FileWriter> m_addrWriter;
-  m4::Tree<Place> m_places;
+
   RelationTagsNode m_nodeRelations;
   RelationTagsWay m_wayRelations;
 
@@ -334,32 +260,25 @@ class OsmToFeatureTranslator
     return params.IsValid();
   }
 
-  void EmitFeatureBase(FeatureBuilder1 & ft, FeatureParams const & params)
+  void EmitFeatureBase(FeatureBuilder1 & ft, FeatureParams const & params) const
   {
     ft.SetParams(params);
     if (ft.PreSerialize())
     {
       string addr;
-      if (m_addrWriter && ftypes::IsBuildingChecker::Instance()(params.m_Types) && ft.FormatFullAddress(addr))
-        m_addrWriter->Write(addr.c_str(), addr.size());
-
-      static uint32_t const placeType = classif().GetTypeByPath({"place"});
-      uint32_t const type = params.FindType(placeType, 1);
-
-      if (type != ftype::GetEmptyValue() && !ft.GetName().empty())
+      if (m_addrWriter && ftypes::IsBuildingChecker::Instance()(params.m_Types) &&
+          ft.FormatFullAddress(addr))
       {
-        m_places.ReplaceEqualInRect(Place(ft, type),
-            [](Place const & p1, Place const & p2) { return p1.IsEqual(p2); },
-            [](Place const & p1, Place const & p2) { return p1.IsBetterThan(p2); });
+        m_addrWriter->Write(addr.c_str(), addr.size());
       }
-      else
-        m_emitter(ft);
+
+      m_emitter(ft);
     }
   }
 
   /// @param[in]  params  Pass by value because it can be modified.
   //@{
-  void EmitPoint(m2::PointD const & pt, FeatureParams params, osm::Id id)
+  void EmitPoint(m2::PointD const & pt, FeatureParams params, osm::Id id) const
   {
     if (feature::RemoveNoDrawableTypes(params.m_Types, feature::GEOM_POINT))
     {
@@ -370,7 +289,7 @@ class OsmToFeatureTranslator
     }
   }
 
-  void EmitLine(FeatureBuilder1 & ft, FeatureParams params, bool isCoastLine)
+  void EmitLine(FeatureBuilder1 & ft, FeatureParams params, bool isCoastLine) const
   {
     if (isCoastLine || feature::RemoveNoDrawableTypes(params.m_Types, feature::GEOM_LINE))
     {
@@ -380,7 +299,7 @@ class OsmToFeatureTranslator
   }
 
   template <class MakeFnT>
-  void EmitArea(FeatureBuilder1 & ft, FeatureParams params, MakeFnT makeFn)
+  void EmitArea(FeatureBuilder1 & ft, FeatureParams params, MakeFnT makeFn) const
   {
     using namespace feature;
 
@@ -570,13 +489,5 @@ public:
   {
     if (!addrFilePath.empty())
       m_addrWriter.reset(new FileWriter(addrFilePath));
-  }
-
-  void Finish()
-  {
-    m_places.ForEach([this] (Place const & p)
-    {
-      m_emitter(p.GetFeature());
-    });
   }
 };
